m68k: m5441x: set mbar to the end of SRAM
Quoting [1]:

***
Even if SRAM is not used, this is required to avoid
inconsistency at runtime. Otherwise, an assertion failed in ld.so during
init process execution:

Inconsistency detected by ld.so: rtld.c: 1290: dl_main: Assertion `_rtld_local._dl_rtld_map.l_libname' failed!
***

As the SRAM is not supported, just set the address to the end of SRAM.

[1]: https://lore.kernel.org/linux-m68k/1436349092-2214-7-git-send-email-yannick.gicquel@gmail.com/

Signed-off-by: Yannick GICQUEL <yannick.gicquel@gmail.com>
Signed-off-by: Jean-Michel Hautbois <jeanmichel.hautbois@yoseli.org>