m68k: coldfire: Enable DMA timer 0 support for M5441x
In order to use high resolution timers, the M5441x can use the DMA
timers. It has four of it, but let's use the first one only, as it is
based on the existing driver for the MCF523x.

While enabling the clock, fix a small typo in the comments too (timers
where starting at 1 and not 0).

With this commit, it is now possible to use high resolution timers, and
have sub-ms resolution in the kernel timestamps.

Signed-off-by: Jean-Michel Hautbois <jeanmichel.hautbois@yoseli.org>