m68k: coldfire: Support resources for UART
In order to use the eDMA channels for UART, the mcf_platform_uart needs
to be changed. Instead of adding another custom member for the
structure, use a resource tree in a platform_device per UART. It then
makes it possible to have a device named like "mcfuart.N" with N the
UART number.

Later, adding the dma channel in the mcf tty driver will also be more
straightfoward.

Signed-off-by: Jean-Michel Hautbois <jeanmichel.hautbois@yoseli.org>