arch: m68k: Disable cache for DMA pool
The global DMA pool is defined in a DMABASE(DMASIZE) zone. We need to
disable cache for this zone, as it is the only way to make it coherent.
In order to do that, change the ACRs definitions:
- ACR0 is still used for IO maps
- ACR1 covers DMABASE->DMASIZE zone and is marked non-cached
- ACR4 covers all the RAM and is marked cacheable.

With the coldfire cache, for overlapping regions, the lower ACR number
takes priority.

Signed-off-by: Jean-Michel Hautbois <jeanmichel.hautbois@yoseli.org>